reference_python/requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
imageio[ffmpeg]>=2.31.0
mss>=9.0.0
